refactor: Add CreatedAt component for displaying formatted dates
This commit adds a new component called `CreatedAt` that is responsible for displaying formatted dates. It takes a `date` prop as input and uses the `moment` library to format the date. The component checks if the date is the same as the current day and displays the relative time using `fromNow()`, otherwise it formats the date as "MMMM Do YYYY". This component can be used to improve the display of timestamps in the application.